Edits + new Stat
-Гост мод теперь при каждом применении шанс 20% на:
Если у игрока ближний бой - немного увеличить размер оружия
Если у игрока дальний бой - прибавить 1 к максимальным патронам
-Добавил стат, который выдается один раз, и если он у тебя есть то ты можешь ломать пули врагов

+++ Assets/Scripts/EnemySpawnSystem/Enemy.cs
@@ -12,8 +12,8 @@ public class Enemy : MonoBehaviour
     public EnemyPool pool;
     private Knockback knockback;
     private Flash flash;
-    private int lowExpThreshold;
-    private int highExpThreshold;
+    private float lowExpThreshold;
+    private float highExpThreshold;
     private bool InCameraView;
     private LightningController lightningController;
     private SpriteRenderer sp;
@@ -25,10 +25,10 @@ public class Enemy : MonoBehaviour
         return InCameraView;
     }
 
-    public void SetLowExpTh(int Num){
+    public void SetLowExpTh(float Num){
         lowExpThreshold = Num;
     }
-    public void SetHighExpTh(int Num){
+    public void SetHighExpTh(float Num){
         highExpThreshold = Num;
     }
